pyproject.toml
[tool.poetry]
name = "todos_app"
version = "0.1.0"
description = ""
authors = ["Amirhossein Faghihifar <a.faghihifar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
PyYAML = "^6.0"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
Flask-RBAC = "^0.5.0"
psycopg2-binary = "^2.9.3"
flask-marshmallow = "^0.14.0"
Flask-JWT-Extended = "^4.3.1"
marshmallow-sqlalchemy = "^0.28.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"